Handle arbitrary code span starts
Previously we hardcoded the start of a code block as |```mermaid|.
But it can have more backticks. Or other symbols really.
Same for the end.

The parser already gives us all the information we need: The span tells
us about the start and end and between a code block start and end can
only be text, which also has a span.

Fixes #24