Handle CR-LF in mermaid
mermaid blocks containing CR LF are returned line by line from the parser instead
of as a complete code chunk for LF only.
We need to expand the span for each chunk we get.
After the closing ticks we replace the remaining \r\n with \n to have a consistent
line ending in the mermaid code at least.